requirements.txt
Flask==2.1.2
torch==1.12.1
torchvision==0.13.1
tensorflow==2.8.0
numpy==1.22.3
Pillow==9.1.1
gunicorn==20.1.0
